// src/core/model/downloadedCase.ts
export interface DownloadedCase {
  id?: number;
  name?: string;
  email?: string;
  mobile?: string;
  caseStudyId?: number;
  caseName?: string;
  type?: string;
  createDate?: Date;
  updateDate?: Date;
  deleteDate?: Date;
  createById?: number;
  updateById?: number;
  deleteById?: number;
  isDelete?: boolean;
}